Refractored model inversion plotter to move most of it up to tvb_fit
